Multitentancy

+++ aspnet-core/src/InventoryManagementSystem.Application/Companies/Dto/CreateCompanyDto.cs
@@ -9,6 +9,6 @@ namespace InventoryManagementSystem.Companies.Dto
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public long TenantId { get; set; }
+        public int TenantId { get; set; }
     }
 }
