+++ aspnet-core/src/InventoryManagementSystem.Application/ProductSale/Dto/CreateProductSaleDto.cs
@@ -12,7 +12,7 @@ namespace InventoryManagementSystem.ProductSales.Dto
         public double SellingRate { get; set; }
         public long ShopProductId { get; set; }
         public int? Quantity { get; set; }
-        public long TenantId { get; set; }
+        public int TenantId { get; set; }
 
     }
 }
